refactored components and tests adding some extra tests, added view boilerplates
