+++ src/components/WPHeader.vue
