--- src/components/WPHeader/WPNavLogo.vue
@@ -1,14 +1,0 @@
-<template>
-  <div class="wp-logo">
-    <img src="@/assets/wplogo-nav.png" alt="wildpulse logo" />
-  </div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped>
-.wp-logo {
-  width: 300px;
-  height: auto;
-}
-</style>
