+++ src/components/__tests__/WPNavLogo.spec.ts
@@ -1,11 +1,19 @@
+/**
+ * @vitest-environment happy-dom
+ */
 import { WPNavLogo } from '@/components'
+import router from '@/router'
 import { mount } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
 
 const PATH_TO_LOGO = '/src/assets/wplogo-nav.png'
 
 describe('wildpulse navlogo', () => {
-  const wrapper = mount(WPNavLogo)
+  const wrapper = mount(WPNavLogo, {
+    global: {
+      plugins: [router]
+    }
+  })
 
   it('has logo image', () => {
     expect(wrapper.find('img').exists()).toBeTruthy()
